Move to library/simulation/neuron
