+++ obi_one/scientific/tasks/simulation_execution/circuit_simulation_execution.py
@@ -6,7 +6,7 @@ import entitysdk
 from entitysdk import models
 
 from obi_one.scientific.library.circuit import Circuit
-from obi_one.scientific.library.simulation.staging import stage_circuit
+from obi_one.scientific.library.simulation.neuron.staging import stage_circuit
 from obi_one.scientific.tasks.simulation_execution.base import (
     SimulationExecutionSingleConfig,
     SimulationExecutionTask,
