+++ obi_one/scientific/tasks/simulation_execution/ion_channel_simulation_execution.py
@@ -8,7 +8,7 @@ from entitysdk.types import AssetLabel
 from obi_one.core.deserialize import deserialize_obi_object_from_json_data
 from obi_one.core.exception import OBIONEError
 from obi_one.scientific.library.circuit import Circuit
-from obi_one.scientific.library.simulation.staging import stage_ion_channel_models_as_circuit
+from obi_one.scientific.library.simulation.neuron.staging import stage_ion_channel_models_as_circuit
 from obi_one.scientific.tasks.generate_simulations.config.neuron.neuron_ion_channel_models import (
     IonChannelModelSimulationSingleConfig,
 )
