+++ tests/obi_one/scientific/tasks/simulation_execution/test_ion_channel_simulation_execution.py
@@ -2,17 +2,17 @@ from unittest.mock import MagicMock, patch
 from uuid import UUID
 
 import pytest
+from obi_one.scientific.library.simulation.schemas import (
+    NeurodamusMechanismBuild,
+    NeurodamusSimulationParameters,
+    SimulationResults,
+)
 
 from obi_one.core.info import Info
 from obi_one.scientific.blocks.ion_channel_model.ion_channel_model import (
     IonChannelModelWithConductance,
 )
 from obi_one.scientific.from_id.ion_channel_model_from_id import IonChannelModelFromID
-from obi_one.scientific.library.simulation.schemas import (
-    NeurodamusMechanismBuild,
-    NeurodamusSimulationParameters,
-    SimulationResults,
-)
 from obi_one.scientific.tasks.generate_simulations.config.neuron.neuron_ion_channel_models import (
     IonChannelModelSimulationScanConfig,
     IonChannelModelSimulationSingleConfig,
